+++ app/controllers/Authenticate.java
@@ -1,22 +1,20 @@
 package controllers;
 
-import be.objectify.deadbolt.core.models.Subject;
 import com.google.inject.Inject;
-import de.qaware.heimdall.Password;
-import de.qaware.heimdall.PasswordFactory;
-import persistence.dao.UserDao;
+import exceptions.AuthException;
 import persistence.entity.AuthorisedUser;
 import play.data.Form;
 import play.libs.F;
 import play.mvc.Controller;
 import play.mvc.Result;
+import services.security.AuthService;
 import views.html.auth.login;
 
 import java.io.IOException;
-import java.util.Optional;
+import java.util.UUID;
 
 public class Authenticate extends Controller {
-  @Inject UserDao userDao;
+  @Inject private AuthService authService;
 
   public F.Promise<Result> authenticate() {
     return F.Promise.promise(() -> {
@@ -25,27 +23,20 @@ public class Authenticate extends Controller {
       if ( loginForm.hasErrors() ) {
         return badRequest(login.render(loginForm));
       }
-
-      final String formPassword = loginForm.get().getPassword();
-      final Optional<Subject> byEmail = userDao.findByEmail(loginForm.get().getEmail());
-      if ( byEmail.isPresent() ) {
-        final AuthorisedUser authorisedUser = (AuthorisedUser) byEmail.get();
-        final Password password = PasswordFactory.create();
-        if ( password.verify(formPassword.toCharArray(), authorisedUser.getPassword()) ) {
-          // Check if the hash uses an old hash
-          if ( password.needsRehash(authorisedUser.getPassword()) ) {
-            // algorithm, insecure parameters, etc.
-            String newHash = password.hash(formPassword.toCharArray());
-            authorisedUser.setPassword(newHash);
-            userDao.upsert("email", authorisedUser.getEmail(), authorisedUser, authorisedUser);
-          }
-          session().clear();
-          session("email", loginForm.get().email);
-          return redirect(routes.Passwords.index());
-        }
+      try {
+        final UUID sessionId = authService.authorise(loginForm.get());
+        session().clear();
+        session("id", sessionId.toString());
+      } catch (AuthException e) {
+        session().clear();
+        loginForm.reject(e.getMessage());
+        return badRequest(login.render(loginForm));
+      } catch (Exception e) {
+        session().clear();
+        loginForm.reject("Could not authenticate.");
+        return badRequest(login.render(loginForm));
       }
-      loginForm.reject("No user found");
-      return badRequest(login.render(loginForm));
+      return redirect(routes.Passwords.index());
     });
   }