+++ app/controllers/Upload.java
@@ -1,29 +1,22 @@
 package controllers;
 
 import be.objectify.deadbolt.java.actions.SubjectPresent;
-import com.google.common.base.Charsets;
-import com.google.common.collect.Lists;
-import com.google.common.net.MediaType;
 import com.google.inject.Inject;
-import persistence.dao.PasswordDao;
-import persistence.entity.Password;
+import exceptions.TypeMismatch;
 import play.data.DynamicForm;
 import play.data.Form;
 import play.libs.F;
 import play.mvc.Controller;
 import play.mvc.Result;
-import utils.keepass.KeypassImporter;
+import services.password.PasswordService;
 import views.html.content.upload;
-
-import java.util.List;
 
 /**
  * Created by Leonard Daume on 23.11.2015.
  */
 @SubjectPresent
 public class Upload extends Controller {
-
-  @Inject PasswordDao passwordDao;
+  @Inject private PasswordService passwordService;
 
   public F.Promise<Result> index() {
     return F.Promise.promise(() -> ok(upload.render(Form.form().bindFromRequest())));
@@ -35,28 +28,21 @@ public class Upload extends Controller {
       play.mvc.Http.MultipartFormData body = request().body().asMultipartFormData();
       play.mvc.Http.MultipartFormData.FilePart csv = body.getFile("csv");
 
-      if ( csv != null ) {
-        String contentType = csv.getContentType();
-        final MediaType mediaType = MediaType.parse(contentType);
-        final List<MediaType> allowedTypes = Lists.newArrayList(MediaType.CSV_UTF_8,
-                                                                MediaType.MICROSOFT_EXCEL,
-                                                                MediaType.OPENDOCUMENT_SPREADSHEET,
-                                                                MediaType.OOXML_SHEET);
-        if ( !allowedTypes.contains(mediaType) && !allowedTypes.contains(mediaType.withCharset(Charsets.UTF_8)) ) {
-          uploadForm.reject("Only the types "
-                            + allowedTypes
-                            + " are allowed.\nBut {"
-                            + mediaType
-                            + "} was uploaded"
-                            + ".");
-          return badRequest(upload.render(uploadForm));
-        }
-        List<Password> passwords = KeypassImporter.fromKeepassCSV(csv.getFile());
-        passwordDao.saveOrUpdateAllCredentials(passwords);
-        return redirect(routes.Passwords.index());
+      if ( csv == null ) {
+        uploadForm.reject("Missing file");
+        return badRequest(upload.render(uploadForm));
       }
-      uploadForm.reject("Missing file");
-      return badRequest(upload.render(uploadForm));
+
+      try {
+        passwordService.importCsv(csv);
+      } catch (TypeMismatch typeMismatch) {
+        uploadForm.reject(typeMismatch.getMessage());
+        return badRequest(upload.render(uploadForm));
+      } catch (Exception e) {
+        uploadForm.reject("Could not import csv.");
+        return badRequest(upload.render(uploadForm));
+      }
+      return redirect(routes.Passwords.index());
     });
   }
 }