+++ app/exceptions/AuthException.java
@@ -1,0 +1,10 @@
+package exceptions;
+
+/**
+ * Created by leonard on 16.02.16.
+ */
+public class AuthException extends Exception {
+  public AuthException(final String message) {
+    super(message);
+  }
+}