+++ app/exceptions/TypeMismatch.java
@@ -1,0 +1,10 @@
+package exceptions;
+
+/**
+ * Created by leonard on 16.02.16.
+ */
+public class TypeMismatch extends Exception {
+  public TypeMismatch(final String message) {
+    super(message);
+  }
+}