+++ app/utils/keepass/KeypassImporter.java
@@ -1,6 +1,9 @@
 package utils.keepass;
 
-import com.google.common.base.Splitter;
+import com.fasterxml.jackson.databind.MappingIterator;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.csv.CsvMapper;
+import com.fasterxml.jackson.dataformat.csv.CsvSchema;
 import com.google.common.collect.Lists;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -8,9 +11,10 @@ import persistence.entity.Password;
 import play.Logger;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Map;
+
+import static org.apache.commons.lang3.StringUtils.trimToEmpty;
 
 /**
  * Created by Leonard Daume on 22.11.2015.
@@ -20,16 +24,23 @@ public class KeypassImporter {
   public static List<Password> fromKeepassCSV(final File csvFile) {
     final List<Password> passwords = Lists.newArrayList();
     try {
-      passwords.addAll(FileUtils.readLines(csvFile).stream().skip(1).map(line -> {
-        List<String> entries = Splitter.on(",").trimResults().splitToList(line);
-        return new Password(stripQuotes(entries, 0),
-                            stripQuotes(entries, 1),
-                            stripQuotes(entries, 2),
-                            stripQuotes(entries, 3),
-                            stripQuotes(entries, 4));
-      }).collect(Collectors.toList()));
-    } catch (IOException e) {
+      ObjectMapper mapper = new CsvMapper();
+      final String csvString = FileUtils.readFileToString(csvFile);
+      final MappingIterator<Map<String, String>> passWordIterator = mapper.readerFor(Map.class)
+                                                                          .with(CsvSchema.emptySchema()
+                                                                                         .withHeader()
+                                                                                         .withEscapeChar('\\'))
+                                                                          .readValues(csvString);
+      passWordIterator.forEachRemaining(passwordMap -> {
+        passwords.add(new Password(trimToEmpty(passwordMap.get("Account")),
+                                   trimToEmpty(passwordMap.get("Login Name")),
+                                   trimToEmpty(passwordMap.get("Password")),
+                                   trimToEmpty(passwordMap.get("Web Site")),
+                                   trimToEmpty(passwordMap.get("Comments"))));
+      });
+    } catch (Exception e) {
       Logger.error("Could not map keepass csv");
+      throw new RuntimeException(e);
     }
     return passwords;
   }