+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+v1.1.0 (2016-0X-XX)
+-------------------
+
+* add or edit passwords with a helping password generator.
+
 v1.0.0 (2016-02-XX)
 -------------------
+
+* Add/Edit/Delete passwords. ✓
  
 * Export options include copy to clipboard, CSV, Excel (XSLX) and PDF files ✓
 
@@ -7,10 +14,9 @@ v1.0.0 (2016-02-XX)
 
 * server side user-password hashing with [Heimdall - Secure Password Hashing](https://github.com/qaware/heimdall) ✓
 
-* Add/Edit/Delete passwords.
- * add or edit passwords with a helping password generator.
+* Session caching with timeout ✓
 
 v0.0.1-SNAPSHOT (2016-02-09) ✓
 ------------------------
 
-* First runnable release of Play-Pass
+* First runnable release of Play-Pass ✓